Added trouble shooting steps for RCTPushNotificationManager.h file not found
It is found that many people had faced issue with  `RCTPushNotificationManager.h file not found`. I too have faced this issue multiple times and had to solve this issue in multiple ways. 

Added troubleshooting section and added various solutions that worked for me under this section and linked one github discussion related to this post.